requirements.txt
streamlit
pycryptodome
PyPDF2
reportlab
